requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
azure-storage-blob==12.19.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
aiofiles>=23.2.1